clean up merge
